Move fetching project contents into a separate function
